+++ Changelog.md
@@ -5,6 +5,7 @@
 + Added `VideoKitCameraManager.StartRunningAsync` method which returns a `Task` for async completion and error handling.
 + Added `VideoKitAudioManager.StartRunningAsync` method which returns a `Task` for async completion and error handling.
 + Added `VideoKitRecorder.StartRecordingAsync` method which returns a `Task` for async completion and error handling.
++ Added `VideoKitRecorder.StopRecordingAsync` method which returns a `Task` for async completion and error handling.
 + Fixed `CameraDevice.StopRunning` method causing app hanging when camera device is unsupported (#126).
 + Refactored `MediaAsset.Caption<T>` method to `MediaAsset.Parse<T>` for structured parsing.
 + Refactored `MediaAsset.Caption` method to `MediaAsset.Transcribe` for performing speech-to-text.